vercel-project/requirements.txt
psycopg2-binary==2.9.9
pydantic>=2.6.4
boto3>=1.34.129
opencv-python>=4.8.0
litellm>=1.73.0
httpx>=0.25.0
PyJWT>=2.8.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
aiofiles>=23.2.1